+++ src/TestSuite/ConsoleIntegrationTestCase.php
@@ -14,6 +14,7 @@
 namespace MeTools\TestSuite;
 
 use Cake\Console\Shell as CakeShell;
+use Cake\Http\BaseApplication;
 use Cake\TestSuite\ConsoleIntegrationTestCase as CakeConsoleIntegrationTestCase;
 use MeTools\Console\Shell;
 use MeTools\TestSuite\Traits\MockTrait;
@@ -35,6 +36,13 @@ abstract class ConsoleIntegrationTestCase extends CakeConsoleIntegrationTestCase
     protected $Shell;
 
     /**
+     * Application instance
+     * @since 2.18.0
+     * @var \Cake\Http\BaseApplication
+     */
+    protected $app;
+
+    /**
      * If `true`, a mock instance of the shell will be created
      * @var bool
      */
@@ -44,11 +52,15 @@ abstract class ConsoleIntegrationTestCase extends CakeConsoleIntegrationTestCase
      * Called before every test method
      * @return void
      * @uses $Shell
+     * @uses $app
      * @uses $autoInitializeClass
      */
     public function setUp()
     {
         parent::setUp();
+
+        $this->app = $this->getMockForAbstractClass(BaseApplication::class, ['']);
+        $this->app->addPlugin('MeTools')->pluginBootstrap();
 
         if (!$this->Shell && $this->autoInitializeClass) {
             $parts = explode('\\', get_class($this));